refactor: interceptor 활성화, theme 생성 및 삭제에 적용
