+++ src/test/java/nextstep/member/MemberE2ETest.java
@@ -1,0 +1,29 @@
+package nextstep.member;
+
+import io.restassured.RestAssured;
+import nextstep.member.dto.MemberRequest;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.test.annotation.DirtiesContext;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+public class MemberE2ETest {
+    @DisplayName("멤버를 생성한다")
+    @Test
+    public void create() {
+        MemberRequest body = new MemberRequest("username", "password", "name", "010-1234-5678", "member");
+        RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(body)
+                .when().post("/members")
+                .then().log().all()
+                .statusCode(HttpStatus.CREATED.value());
+    }
+}
